Hunter v1.9.18 (#6)
* v1.9.17

* fix vulkanmemalloc

* add missing includes

Should fix #3633

* fix rendertarget sample app

we can't use a projection with infinite far as a custom
projection matrix, because this is not suited for culling.

recently a new version of setCustomProjection was added which allows
to specify both the rendering and culling projections.

* gltfio: Refactor the tangents job.

This cleans up the glTF tangents generator and allows it to be used
in other parts of the library.

* gltfio: Refactor Animator.

This does not change any behavior, it just prepares for the upcoming
MorphHelper class.

* Fix typo with Metal createRenderTarget

* Fix RenderTarget slice bug (#3656)

* gltfio: Add internal MorphHelper, enable up to 255 targets. (#3661)

* gltfio: Add internal MorphHelper, enable up to 255 targets.

Previously, gltfio could not animate models with more than 4 morph
targets. Now we can handle up to 255 targets, as long as no more than
4 targets are ever used simultaneously. This allows us to handle
many more models, such as those exported by Modo, and the horse model
in #1852.

We do this by picking the highest 4 weights. The indices for these
weights are called the "primary indices", and we construct a new
VertexBuffer for each unique set of primary indices.

In the near future this will be made more efficient by adding a new
VertexBuffer API to the Filament core.

Fixes #1487.

* Code review fixups.

* gltfio: Add support for data:// in image URI's. (#3662)

* Fix blanking windows not being drawn into on macOS (#3616)

* Fix blanking windows not being drawn into on macOS

* Refactored to create swap chain info object; added fixes for resizing; added sample.

* CocoaGLSwapChain should inherit from Platform::SwapChain

* Link to SDL directly for multiple_windows example

* SDL apparently redefines main to SDL_main and requires C++ linkage

* Fix main to have SDL-approved signature

* DoF: minor optimizations

we can refactor the code to get rid off the call to mergeRings() for 
the first ring, and we don't need to initialize the current bucket
either.

* DoF: always apply the correct sample weight

before we were not doing that because it caused some aliasing
artifacts with mipmapping, however it turns out that there are artifacts
anyways, and using the proper weights actually reduces the most severe
ones.

the artifacts are due -- it seems -- to the low density of the kernel.

* DoF: fix a bug that caused black dots around the fast tiles

this would happen only when a fast tile had a CoC < 2, because it
would then have an opacity, but we were ignoring it. It's color then
didn't match the color of the surrounding layer and the median filter
would spread a pixel from the surrounding layer into the fast tile.

we simply fix this by (1) generating the alpha/color for fast tiles,
which is easy to do and on the combine pass, skip the fast tile 
combine when the tile has a an opacity.

* Add deprecation annotation to Camera.

* Update build scripts to match latest AGP and Kotlin

* Update RELEASE_NOTES for 1.9.17

* Bump version to 1.9.18

* RenderTarget: fix NPE when depth is not present.

This fixes a recent regression that would occur when a RenderTarget
does not have a depth attachment.

* Update RELEASE_NOTES for 1.9.18

* Update version

* update hunter repos

* fix mac build

Co-authored-by: Eli Semory <elisemory@snap.com>
Co-authored-by: Mathias Agopian <mathias@google.com>
Co-authored-by: Philip Rideout <philiprideout@gmail.com>
Co-authored-by: Benjamin Doherty <bendoherty@google.com>
Co-authored-by: prewettg <prewettg@gmail.com>
Co-authored-by: Romain Guy <romainguy@curious-creature.com>